Add workflow-ui using webpack
Fixing routing in api-gateway, changing Dockerfile. This is work in
progress i.e. this is a first draft, not a fully working version

Test Plan:
Create tenant 'fb-test', user 'fbuser' and go to http://fb-test.localtest.me:5000/wf
The user interface should be available

Known issues:
User interface is not yet connected to the proxy, no data will be
available. The interface also does not respond to roles (admin or user).