Fixing builder redirect and task name
Issue: NH-577
Test Plan: as a non-admin when you go to the builder URL you will be
redirected back to the defs screen. In the builder in Tasks menu the
individual tasks won't have a (technical) prefix "GLOBAL".

When user tries to go to the builder URL in the service menu
he should be redirected back to the initial screen. Tasks don't have
a prefix which is a byproduct of using workflow proxy.